reuse allspecificsitesofffield component
